+++ Source/Mosa.CoolWorld/HAL/HardwareAbstraction.cs
@@ -59,7 +59,7 @@ namespace Mosa.CoolWorld.HAL
 		/// Processes the interrupt.
 		/// </summary>
 		/// <param name="irq">The irq.</param>
-		public void ProcessInterrupt(byte irq, uint error)
+		public void ProcessInterrupt(byte irq, byte error)
 		{
 			Mosa.DeviceSystem.HAL.ProcessInterrupt(irq, error);
 		}
